input derivatives only calculated once
Adjusted function to use library_config dictionary to reciev extra parameters and then used this dictionary to store half of the values calculated so that they do not need to be computed every time. The input derivatives are independant of the NN and so don't need to be recalculated every epoch.